requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dateutil>=2.8.2
isodate>=0.6.1
tqdm>=4.66.0
regex>=2023.10.3
playwright>=1.40.0
